Translate VLAN IDs after replication
Depending on the network, you might want to use different
VLAN IDs for each of the replicated stream.

This was not possible before, because the replication happens
AFTER the XDP program is finished due to the
bpf_redirect_map(tx, 0, BPF_F_BROADCAST | BPF_F_EXCLUDE_INGRESS)

Therefore, attach a second XDP program to be executed
after the XDP_REDIRECT was performed that only changes
the VLAN ID.

This also allows to simplify the configuration by
only changing the VLAN IDs once per FRER function (replication
or elimination) and not twice (before and after).

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>