remove hardcoded config
With this commit, the hardcoded config was removed,
so we must use command-line arguments to create the
config. It seems this is a step back, but this allows
us to use different FRER instances on multiple devices
in various scenarios.

There are many command-line arguments you can use to
configure the XDPFRER. You have to use `-m` to select
mode, which can be replication or elimination. With
`-i` you can set ingress interfaces with VLAN ID-s.
On the other hand, `-e` sets egress interfaces. `-q`
hides the output. When you have multiple replication
and elimination functions in the network, occasionally,
XDPFRER doesn't need to remove the R-tag at elimination.
With `-n` you can tell XDPFRER not to remove or add an R-tag.